[miopen] Fix Clang Format Check in CI  (#2053)
## Motivation
We've noticed some inconsistencies with our code formatting where our
local make formats would touch files that were not part of our PR.
However, CI would not catch these changes. It turns out that the Clang
format check was silently failing on the CI due to an error since the
migration to the monorepo.

## Technical Details
The error log indicates that the Clang format check was failing:

```
2025-10-08  01:36:28  + find /projects/miopen -iname *.h -o -iname *.hpp -o -iname *.cpp -o -iname *.h.in -o -iname *.hpp.in -o -iname *.cpp.in -o -iname *.cl
2025-10-08  01:36:28  + grep -v -E (build/)|(install/)|(fin/)
2025-10-08  01:36:28  + xargs -n 1 -P 1 -I{} -t sh -c clang-format-12 -style=file {} | diff - {}
2025-10-08  01:36:28  xargs: warning: options --max-args and --replace/-I/-i are mutually exclusive, ignoring previous --max-args value
2025-10-08  01:36:28  find: '/projects/miopen': No such file or directory
```
The Clang format check was implemented using a custom command in the
Jenkinsfile. However, since we already have a `make check_format` target
that reliably detects formatting issues across the repository, we are
replacing the custom command with this make target.

This also addresses a minor bug which would generate an incorrect
directory name due to a duplicate backslash in the repoDir name.

## Test Plan

1) Modify the Jenkinsfile to replace the custom Clang format check
command with make check_format.
2) Ensure that the make check_format target runs successfully and
catches any formatting issues.
3) Verify that the CI pipeline fails if there are any formatting issues
in the codebase.

## Test Result

1) The make `check_format` target was integrated into the Jenkinsfile.
2) The CI pipeline now correctly identifies and fails on any formatting
issues.
3) The minor bug related to the duplicate backslash in the repoDir name
has been resolved.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: JonathanLichtnerAMD <195780826+JonathanLichtnerAMD@users.noreply.github.com>