[Hipblaslt] Add rocm 7.0.2 preset (#2082)
## Motivation

We are using the `preset` flag to build hipblaslt and currently 7.0.2 is
not an option. Need to add this to be able to build on 7.0.2

## Technical Details

Allow using the following build command
```
cmake --preset rocm-7.0.2 -DGPU_TARGETS="${PYTORCH_ROCM_ARCH}" -B build -S. \
    && cmake --build build --target package --parallel 
```

## Test Plan

Built image with this change and see our integration test succeeding.

## Test Result

All test passed.

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Signed-off-by: Gene Der Su <e870252314@gmail.com>