rocsparse_sptrsv - rocsparse_sptrsm (#1331)
New routines rocsparse_sptrsv and rocsparse_sptrsm.
Refactoring of the internal backend.

## Technical Details

The refactoring of the backend addresses the feature that
rocsparse_sptrsv has a descriptor with possible shared data with
rocsparse_mat_info. Data structures are introduced to simplify and
structure the collection of trm_info.

## Test Plan

All tests, including newly introduced ones, are passing.