[Ex CI] Prevent rocm-libraires-superbuild from being cancelled (#2248)
## Motivation
The Azure pipelin for rocm-libraires-superbuild was being cancelled.
<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details
The Trigger file for superbuild triggers a pipeline build off of a new
PR. However the Azure ci dispatcher will check for any running builds
for a PR and cancel those builds. This caused superbuild to be cancelled
just after it had started. I added an exception in the azure ci
dispatcher to exclude the rocm-libraires-superbuild pipeline as the
trigger for it should remain as is.
<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.