[MIOpen] Disable AI heuristics for 3D CK WRW solver (#2290)
The AI heuristics were picking an invalid value of splitk in some cases
for the 3D CK WRW solver.

## Motivation

This PR selectively reverts just the portion of the 3D heuristics that
were causing an issue.

## Test Plan

Reproduced issue with a failing shape using MIOpenDriver.

## Test Result

Verified that this shape passes after these heuristics were disabled.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.