[Stream HPC] Update sorting verification for tests (#498)
There is
[`device_sort_check`](https://github.com/ROCm/rocm-libraries/blob/develop/projects/rocprim/test/rocprim/test_utils_sort_checker.hpp)
utilities that can verify on device if the data in a `device_ptr` is
sorted, under some conditions it would be faster than the sort checking
on host, this PR roughly compares the time taken by device and host
check, and replace host checking when device checking is faster.

The attachment is the benchmarking logs from the comparing.

[timespent.zip](https://github.com/user-attachments/files/21060175/timespent.zip)

---------

Co-authored-by: Sander Bos <sander@streamhpc.com>