[MIOpen] Port im2col and col2im to HIP (#1882)
## Motivation

This PR is part of the effort to port remaining OpenCL kernels in MIOpen
to HIP. This pull request focuses on porting the `im2d2col`, `im3d2col`,
`col2im2d` and `col2im3d` kernels.


## Technical Details

Ported the following kernels to HIP:

* `Col2Im2dU`
* `Col2Im3dU`
* `Im2d2Col_v2`
* `Im3d2Col`


## Test Plan

Changes can be verified using the verification in the `MIOpenDriver`, or
using existing tests that invoke the `GemmBwdRest` or `GemmFwdRest`
solvers such as the unit tests with the same name, or `test_conv_2d`,
`test_conv_3d` or by the google tests. Since no new functionality is
added here, the existing tests should be sufficient to cover this code
change