build(rocprim): fix offload compress detection not working (#2313)
## Motivation

We've noticed CI failures due to linker errors.

## Technical Details

When building tests on multiple architectures, linking sometimes fails
due to big/fat kernels bloating the code object file. Offload compress
is supposed to help with that. But this broke in the recent updates.

## Test Plan

When compiling on ROCm 7.x, offload compress detection should not fail
on standard ROCm setups.

## Test Result

Works on my machine (and our internal CI).

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.