[CI] Fix "latest" update failure in update-docs.yml (#2437)
## Motivation

The pipeline triggers version rebuild upon push action to keep our docs
synchronized. For latest docs branch, it not only needs to trigger
rebuild for docs/x.y.z, but also needs to trigger "latest". The current
implementation ignores "latest". This PR aims to fix this issue.

## Technical Details

Extracts latest version from rocm-docs-core new_data branch. And compare
to check if the updated branch is latest. If so, trigger the
corresponding project latest version to build.

## Test Plan

Local test:
Define `branch_name="docs/7.1.0"`, `project="rocblas"`. Execute the
added code and observe if both "docs/7.1.0" and "latest" on RTD are
triggered for rocblas.

## Test Result

`latest` and `docs/7.1.0` are triggered to rebuild as expected.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.