Cherry-pick build fixes for rocThrust + libhipcxx for ROCm 7.1.1 (#2465)
Do not merge until PM approval given. 

## Motivation

We have discovered an issue with the rocThrust version that comes with
ROCm 7.1. In particular, once libhipcxx is installed on a system,
rocThrust will include libhipcxx headers. Unfortunately, the version of
rocThrust is not aligned with the 2.7.0 libhipcxx version and seems to
use CCCL includes from >=2.8.x. Hence, any include of rocThrust will
fail once libhipcxx is installed on a ROCm 7.1 system. Fixing this in
rocThrust should be rather easy by disabling the hip/std includes for an
incompatible libhipcxx version.

## Technical Details

This PR combines #2430, #1769, and #1770

## Test Plan

Running CI tests now.

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: Yung-sheng Tu <112800063+yungshengtu@users.noreply.github.com>
Co-authored-by: Nara <nara@streamhpc.com>
Co-authored-by: Saiyang Zhang <41819448+Saiyang-Zhang@users.noreply.github.com>