[rocPRIM] device_radix_sort_onesweep performance optimization (#2056)
This is a performance optimization for device_radix_sort_onesweep. That
results in great performance increases for the MI308 and good
performance increases on the MI300.

The amount of LDS usages was reduced to get better occupancy. We are
reusing LDS memory. This optimization is only done on types larger then
short.

Other optimization the double2 type now performs as good as the
custom_type with two doubles by using a value wrapper.
Also for other GPUs besides gfx942, an optimization is done to reduce
the amount of branching, which was necessary to remove regressions on
other gpus.

The device_radix_sort tests should compile and pass.

device_radix_sort compiles and passes.

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: Nara Prasetya <nara@streamhpc.com>
Co-authored-by: Wayne Franz <wayfranz@amd.com>