rocfft: fix callbacks for single-proc multi-GPU transforms
## Motivation

Fix load/store callbacks on single-process multi-GPU transforms.

## Technical Details

Callbacks were not being attached to the correct TreeNodes during
multi-GPU plan execution. Even though callbacks are only known at
execute time, we can piggyback on the load/store ops infrastructure for
extra operations we want to do that are known at plan creation time.

Modified the load/store ops to be a std::optional on plans. If the
optional load/store op is present, then that means this node in the plan
is doing the first load from global memory or last store to global
memory for the FFT. The load/store op will std::nullopt for internal
nodes of a multi-kernel FFT, or for communication steps.

At execution time, we can then look at presence/absence of the
load/store op to know if load/store callbacks should be applied to that
node of the transform.

We also need to build up an internal device->callback map, so that we
can know exactly which callback to invoke for the device that this node
executes on. This can only be constructed at a time when we have both
the plan description (which knows the field/brick layout, and
specifically the ordering of bricks and the devices they reside on) and
an execution info (which has the callbacks supplied by the user).

## Test Plan

Enabled callback test cases in rocFFT and hipFFT for single-proc
multi-GPU transforms.

## Test Result

Tests pass.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: Ayala <alaayala@amd.com>