[hipsparse] Add cuda deprecation notices to the doxygen documentation (#2548)
## Motivation

1. Add documentation indicating each routines support when using cuda
backends
2. Add other improvements like indicating what parameters must be
positive etc.
3. Fix other minor typos

## Technical Details

Because hipSPARSE is a wrapper library that supports both using cusparse
and rocsparse backends, many routines that hipSPARSE has are now
deprecated in cusparse. This means that when using the cusparse backend,
these functions are no longer supported. To let users know what versions
of cusparse a routine is supported, I have added deprecation warnings
indicating when a routine has been deprecated in cusparse and when it
has been removed in cusparse.