CustomSchedule.py schedules are functions (#2637)
- Existing schedules are refactored to callable functions
 - Add decorator to safely create "isTN/NT/NN/TT" functions
 - All functions (for now) stay in CustomSchedule.py

We are about to land a significant number of custom schedules for
various tile sizes and transpose orientations. By using the convention
"a schedule is described in a function" we can reasonably scale the code
out, instead of a chain of if/elif blocks that spans hundreds or
thousands of lines. Initially we keep the existing schedules in
functions inside CustomSchedule.py, but we can refactor as needed.

 - Existing schedules are refactored to callable functions
 - Add decorator to safely create "isTN/NT/NN/TT" functions
 - All functions (for now) stay in CustomSchedule.py

TENSILELITE_CLIENT_ARGS="--build-type Debug --gpu-targets gfx950" tox -e
py3 -- Tensile/Tests -k custom_mainloop_scheduling.yaml

PASS

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.