[hipDNN] Add TheRock CI for MIOpen plugin project (#2309)
## Motivation

Add TheRock CI pre-submit checks for MIOpen plugin project.
Now that https://github.com/ROCm/TheRock/pull/1811 is merged we can add
the pre-submit checks for MIOpen plugin.

We are working on Windows support for hipDNN & MIOpen plugin.
Once that work is complete we can do a follow-up to add MIOpen plugin as
a pre-submit check for MIOpen changes.
(We can't currently do that since MIOpen will get skipped on Windows if
we try to enable this right now).