[rocSOLVER] Update rocsolver tests to measure median time (#2455)
## Motivation
The mean of gpu runtimes is too sensitive to outliers. This makes the
results of `rocsolver-bench` inconsistent between runs on sensitive
hardware. This PR changes the tests to compute the median instead of the
mean.

## Technical Details
Added `rocsolver_timer` class to control the behavior of collecting and
combining execution times in the tests.

## Test Plan
Compare before and after benchmark results (sytrd).

## Test Result
No significant difference between mean and median. Therefore, the new
time captured is representative of what we had prior.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.