[rocRoller] Fix TopologicalCompare (#2593)
## Motivation

This is the first PR from my experimental branch. This is refactoring
TopologicalCompare to take a pointer or reference to a KernelGraph
instead of a `std::shared_ptr` since generally we don't have a
`std::shared_ptr` and many of the uses were creating copies of the
`KernelGraph` in order to sort a subset of the nodes.

Also added a `COperationWithBody` concept which matches any type of
Operation that can have a Body edge. Added a ticket to make a
GraphConstraint to check for legal types of edges.

## Testing

Added a new test for the TopologicalCompare functor.