Revert bad logic - Low Offset Overflow  (#2472) (#2647)
## Motivation
In response to a ticket debug.

The problem was the low 32-bits of the read address were being
incremented when moving to the next tile, but the high 32-bits were not.
This could cause a problem if the workspace buffer was allocated with an
address close to the 32-bit boundary - it can create a scenario where
incrementing to the next tile causes the low 32-bits to wrap to 0, and
the carry was not being handled correctly so the read address would be
out of bounds before the beginning of the buffer.

Reverting bad logic from
https://github.com/ROCm/rocm-libraries/pull/1080/

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

Co-authored-by: mahmoodw <44450175+mahmoodw@users.noreply.github.com>
Co-authored-by: mahmoodw <wmahmood@amd.com>