[miopen] Move non-critical HW to own nightly (#2516)
## Motivation
Due to the low availability of hardware on our CI, which is limited to a
single node for Strix Halo, we decided to move the strix halo and gfx908
to their own "Non-Critical HW Nightly"

## Technical Details

- Moved the two devices and their corresponding tests to "Non-Critical
HW Nightly"
- These tests are scheduled to run 1 hour after our regular nightlies.
- Due to jenkins limitations on opcode size, could not maintain
duplicates of these tests within the main test suite.
- Addressed inconsistencies in test names in the pipeline.

## Test Plan
- Verify that gfx908 and strix halo are moved to their own nightlies by
manually triggering a nightly run in a feature branch.

## Test Results
- Verified both full tests and manually triggering nightlies and they
all pass the tests as expected.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.