[hipblaslt] Remove SWaitCnt(dscnt=0) from CMS MAINLOOP when using a single codepath (#2730)
## Motivation

When using a single code path with CMS, the pre-loop wait on LR is
included within the loop, which can lead to unnecessary waits. To avoid
this, this PR adds a label just before calling the MAINLOOP macro.

This issue affected only the CMS implementation with a single code path.
The codebase contains only one such case, and an SWaitCnt(dscnt=0) is
already manually performed at the end of the loop.