Detect if kernel launch would overflow hip grid limits (#2747)
## Motivation

The hip launch grid limit is 2^32 work items. When running algo method
all on a large problem, you can end up with a small MT kernel that
results in a large number of tiles that might overflow hip's limits.

## Technical Details

This change adds a new predicate that detects if a solution would
overflow hip's launch limits. When using algo method all,
isSolutionSupported will now accurately indicate that the kernel will
not be able to solve the problem due to exceeding grid limits.

## Test Plan

Fixes the affected test case locally. A new test case to exercise the
feature is added with this PR, but is currently disable since it
requires a lot of memory, and a little longer to run that the average
test case. Need reviewer comments on whether this test case should be
enabled on CI or not.

## Test Result

Local tests for the issue passed. Awaiting CI results.

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.