[MIOpen] Porting BatchNormBwdPerAct kernels to HIP (#2374)
## Motivation

This PR is part of the effort to port the remaining OpenCL kernels in
MIOpen to HIP. This pull request focuses on porting the
`BatchNormBwdPerAct` kernels.

## Technical Details

- Ported the following kernels to HIP:
  - `MIOpenBatchNormBwdPerActivation`
  - `MIOpenBatchNormBwdPerActivationSaved`
- Updated data types to use definitions from the latest `float_types`
header.
 - Adjusted relevant solvers to align with kernel modifications.
- Corrected a few syntax issues in `driver/bn_driver.hpp` in debug mode.
- Modified `driver/bn_driver.hpp` & `test/fusionHost.hpp` to run tests
correctly in the case of `BatchNormBwdPerActivation` kernel.

## Test Plan

Changes can be verified using `test_bn_bwd`, which tests only the
`BatchNormBwdPerActivationSaved` kernel (since mean and variance are
always provided as inputs).

To test both kernels individually, use `MIOpenDriver` with options:

- `--back 1 --save 0` -> `MIOpenBatchNormBwdPerActivation`
- `--back 1 --save 1` -> `MIOpenBatchNormBwdPerActivationSaved`

Tests can be run for `fp32`, `fp16`, and `bfp16` data types by adjusting
the MIOpenDriver input parameters.

## Test Result

Tests pass on gfx908, gfx90a, and gfx942.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.