[rocprim] add code coverage process (#2507)
## Motivation

The current code coverage process does not work with clang compilations.
We want to upload code coverage reports to codecov.io and therefore we
need a working `coverage` cmake target.

## Technical Details

A couple of important details:
1. Because rocprim is header only, we have to build up a string of
cmdline arguments to pass to llvm-cov. See [lines
48-51](https://github.com/ROCm/rocm-libraries/pull/2507/files#diff-0748f04a03e7a50b1e385d2c834b83194b356aefdf638c3e874db0be2dddae38R48-R51)
of projects/rocprim/test/CMakeLists.txt
2. We also need to embed the `LLVM_PROFILE_FILE` environment variable in
the test cases because ctest will not propogate the variable forward to
the testcases

## Test Plan

A comment on this PR will be made with the coverage report for rocprim.