rocfft: fix C2C plan creation failure with non-contiguous strides (#2817)
## Motivation

Fix some multi-dimensional plan creation failures with non-contiguous
strides.

## Technical Details

Part of FFT plan creation involves judging whether the output of a step
of the plan will fit in an available device buffer. This judgement was
not correctly considering the case when the output and buffer have
non-contiguous strides.

## Test Plan

Enabled some existing-but-disabled test cases to cover this situation.

## Test Result

The tests pass.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.