Make unit tests runnable via a tox environment (#2765)
- Using the tox environment "unit", you can quickly call unit tests once
you have a build of tensile-client already available.

- Add documentation to the README advertising this fact.


## Motivation

Our integration testing (via the YAML files) is a comprehensive test of
the codegen capabilities of tensilelite. We also need to be able to test
the Python components separately. Codegen testing (because it ends up
calling the actual tensilelite-client) can run for 15 minutes up to an
hour or more. Unit testing of Python code can be done in 5-20 seconds,
permitting faster development iteration cycles.

## Technical Details

- In 5ee394f38494b836a80b0fe1c2d837b608a8b35c the Sparse key was removed
from the returned value in matrixInstructionToMIParameters but this test
was never updated. Removing from the test allows the test to pass again.


## Test Plan

 tox -e unit -- Tensile/Tests/unit

## Test Result

29 passed, 1 xfailed in 18.54s

(the one expected failure can be addressed in a follow up PR)

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.