[rocWMMA] Suppress printing of extra information in CTest test suites (#2853)
## Motivation

The massive logs generated from tests make it difficult to determine
what is going on when CI fails, and may potentially cause slowdowns when
multiple tests are running and writing their logs to disk (rocWMMA has
millions of tests).

## Technical Details

Setting `--omit 5` will suppress extra output for skipped and passed
tests.

## Test Plan

Ran with ctest, and checked that output is reduced for passing and
skipped tests.

## Test Result

Reduces the number of lines in test output for passed tests by ~1/3.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.