[rocRoller] Allow UnrollLoops to do K-unrolling for StreamK kernels (#2189)
## Motivation

Support unrolling the accumulation loop in StreamK kernels. 

## Technical Details

In StreamK kernels, there are two nested loops: an outer "streaming"
loop, and the (usual) inner "accumulation loop". The accumulation loop
is the "K loop" and we to be able to unroll it (allowing prefetching
etc).

In the coordinate transform graph, StreamK kernels have side-by-side
ForLoop coordinates: one for the streaming loop, and another for the
accumulation loop. The new unroll coordinate will be placed beside the
accumulation loop coordinate.

Unfortunately, finding and placing strides for the new unroll coordinate
is tricky: and therefore a new edge "Identify" is added, and is used to
essentially give hints as to where strides associated with the new
unroll should be placed.


Next, the loop conditions of StreamK kernels has been changed so that
they look like "iterator < max-value". That is, the top node in the
expression is a "LessThan". This allows the existing logic around adding
tail loops to function as before.

To facilitate this, we pre-compute the number of accumulation tiles that
need to be processed and save them in a register. This becomes the new
"max-value" in the accumulation loop condition.

## Test Plan

The StreamK tests in GEMMTest have been parameterised. A new parameter
for the amount of unrolling has been added.

New unit tests for some of the "for loop" utilities have been added.