rocFFT and hipFFT fftw3 library find fix for non Release builds (#2828)
rocFFT and hipFFT failed to find the fftw3_libs
setting the variable to NOTFOUND when the build
type was RelWithDebInfo instead of Release.

Reason is that the property names IMPORTED_LOCATION_RELEASE and
IMPORTED_IMPLIB_RELEASE that FFTW3-cmake files provides worked only for
the cmake release builds.

Last word in the property name depends
from the cmake build type written with uppercase alphabets. For
RelWithDebInfo builds for example the property name is
IMPORTED_LOCATION_RELWITHDEBINFO instead of being
IMPORTED_LOCATION_RELEASE.

fixes: https://github.com/ROCm/TheRock/issues/2242

## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

Fixes the finding of fftw3 libraries in therock builds when the build
type is for example RelWithDebInfo instead of Release.

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

Tested by fixing therock build that failed earlier for finding the fftw3
libraries
for rocFFT and hipFFT when the build type was RelWithDebInfo.

## Test Result

Build passes now both for different cmake build types and provides the
rocfft-test and hipfft-test applications.

## Submission Checklist

- [X ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

Signed-off-by: Mika Laitio <mika.laitio@amd.com>