rocFFT: fix C2R plan creation failure with 2D_SINGLE and non-default strides
## Motivation

Fix potential failure to create a real inverse plan with non-default
strides, when the plan would use a 2D_SINGLE kernel for the fastest
dimensions.

## Technical Details

Due to a historical deficiency in the 2D_SINGLE code generator, when we
fuse c2r pre-processing we swap the fastest two dimensions for that node
in the plan.

This made it easier to perform the operations in the right order:
- transform slow dimension
- c2r pre-process
- transform fast dimension

But swapping the dimensions in the node confuses buffer assignment,
because it would want to compensate for the real-complex length change
along the wrong dimension.

So instead, fix the generator to be able to maintain the correct order
of
operations in this case, and remove the dimension swapping.

## Test Plan

Enabled some existing-but-disabled test cases to cover this situation.

## Test Result

The tests pass.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.