[hipblast][CMS] Add pytests for validator (#3043)
## Motivation

To validate the schedules already included they must be generated. The
current best way of doing that is by running the tox test using
`custom_mainloop_scheudling.yaml`, but this take >10 minutes and exits
early on the first test.

## Technical Details

Introduce a pytest for each existing schedule which creates and runs the
existing validation passes on each currently implemented schedule.

## Test Plan

This is adding more tests. Running existing tests will cover the same
code as these tests, but will take longer. The yaml tests are taken as
the source of truth.

## Test Result

Passing.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>