[rocRoller] Support loading pre-swizzled scales using direct-to-lds (#1850)
## Motivation

We want to be able to load pre-swizzled scales using direct-to-lds
instructions.

## Technical Details

This PR is mostly mechanical: changing boolean "loadLDS" to load-paths
for the scales; and then has a few fixes for compute-index (getting the
correct dimensions to use for address calculations) and swizzle-scale
(connecting everything properly).

Unfortunately the length of the kernel names becomes too long (both file
name and kernel name). This PR has a small hack to use a hash of the
solution parameters.

## Test Plan

Updated existing tests to add the BufferToLDS load path.

Updated a run in the suites that uses different swizzle-tile-sizes to
use BufferToLDS load path as well.

---------

Co-authored-by: Joao P. L. de Carvalho <joao.carvalho@amd.com>