Added find_libblis function to rocBLAS/clients CMakeLists.txt (#2126)
## Motivation

I was building rocBLAS linked with BLIS and saw that part of the logic
for searching for AOCL was hardcoded with respect to different versions.
Specifically, I have the newest, 5.1.0 installed and the current logic
doesn't account for that.

## Technical Details

I wrote a function to replace the hardcoded logic by using file glob and
regex patterns.

## Test Plan

Tested locally with my 5.1.0 version.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: Torre Zuk <42548444+TorreZuk@users.noreply.github.com>