[hipblaslt] Verification that global reads are not not early (#2943)
With current CMS (custom mainloop scheduling) before global reads start
(GRA/GRB), we need

1) s_waitcnt to ensure that (this) wave has completed all local reads
(LRA0/LRB0), and then
2) s_barrier to ensure all waves in (this) workgroup have have completed
all local reads.

This PR algorithmically checks custom schedules to confirm this is the
case.

Please see the tests to see what the error messages are when 1 and/or 2
are not satisfied.

Follow-up work:

The logic can be simplified, and it should be unified with the schedule
based approach of @Alex-Vasile. Binning by vmfma indices is not needed
-- a representation which is free of vmfmas will be cleaner for this
specific validator.

Signed-off-by: James Newling <james.newling@gmail.com>