[hipblaslt] [CMS] Validator for GRs being finished before LR1s start (#3168)
## Motivation

It is easy to incorrectly start LR1s before all of the GRs are
guaranteed to finish. Errors from this are non-deterministic since they
rely on memory stalls and race conditions.

This is especially true when using `SwapGlobalReadOrder` or giving the
GRs less than 1 whole iteration to complete.

## Technical Details

Similar approach to previous validator. Schedule insturctions into a
linear timeline.
1. Schedule GRs, SWaits, SBarriers, and LR1s. **Assume LR1s are valid,
rely on other pass to guarantee this.**
2. Apply effect of Swaits and SBarriers onto GRs
3. Ensure that between when the last GR is issued and the first
corresponding LR1 is issued that:
   1. There is an SWait that guarantees the GRs to finish.
   2. There is an SBarrier between that SWait and the first LR1. 

Handles `SwapGlobalReadOrder`.

## Test Plan

1. Add extra tests.
5. Run existing tests.

## Test Result

1. New tests all pass.
2. Several existing schedules fail. Separate issues opened to track
them:
  - #3181
  - #3182
  - #3183

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>