Disable conv Gemm* solver VRAM limit (#2393)
## Motivation

<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->
Address issue #1329 by allowing the `GemmFwdRest` solution to be used
instead of the naive solver. The `WORKAROUND_MLOPEN_ISSUE_1430`
addresses a problem that originally affected Vega 20 architecture which
is old at this point, and was related to pointer issues in rocBLAS that
have since [been
resolved](https://github.com/ROCm/MIOpen/issues/148#issuecomment-2427619100).

Also implements an option to use verification with GPU reference in the
test framework.

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->
An optional environment variable is added that allows disabling the
maximum workspace size limit in the GEMM solutions, effectively raising
the maximum workspace size from 7.2 GB to the amount of global memory
available.

Since this updates the solver applicability, the system databases had to
be updated to account for this. The databases have been cleaned with
`MIOPEN_DBSYNC_CLEAN=1`, which removes the problematic solver from the
fdb. Only the problematic shape-solver combinations are removed, for
MI300X this was 111 shapes affected out of >200k so database performance
impact should be very small.

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->
Running MIOpenDriver for the shapes specified in #1329, and seeing that
the `GemmFwdRest` solution is chosen.

Implemented tests that use more than 7.2 GB VRAM to ensure the
functionality still works. The tests use GPU reference validation which
reduces runtime from ~60 s per test to ~5 s per test.

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.