[hipsparse] Add yaml testing to csrcolor and fix CHECK_HIP_ERROR macro (#3058)
## Motivation

Adds yaml file testing and additional tests to csrcolor. Also corrects
`CHECK_HIP_ERROR` macro. This PR also has the `testing_<routine_name>`
functions return `void` similar to rocsparse and has the
`CHECK_HIP_ERROR` and `CHECK_HIPSPARSE_ERROR` call `ASSERT_EQ` as
appropriate (similar to rocsparse).

## Technical Details

The `CHECK_HIP_ERROR` macro needs to cache the passed in error: `auto
error = ERROR;` so that if `error != hipSuccess`
and we enter the if statement, we do not re-call the HIP API when
reporting the error. In addition we need to make the `CHECK_HIP_ERROR`
and `CHECK_HIPSPARSE_ERROR` call `ASSERT_EQ` when errors are encountered
in order for the test to properly fail.