[hipsparse] Only compile tests for supported routines (#3041)
## Motivation

## Technical Details

Currently we compile the tests for all routines regardless of backend
used and regardless of whether that backend even supports the routine.
If the routine is not supported, we use preprocessor macros to exclude
the actual test code within the source file. For example, if using the
cusparse 12.8 backend, we compile the level1 routine test translation
units (i.e. `test_axpyi.cpp`, `test_gthr.cpp`, etc) even though cusparse
12.8 does not support any of the level1 routines (they were all
deprecated and then removed). Instead, we use the preprocessor +
`CUDART_VERSION` to selectively remove code so that compilation
completes. This is messy and error prone and it would be much nicer to
remove these preprocessor defines and instead make the act of
compilation itself determine if any errors would occur if the routine
has been removed. With this PR, if the tests compile, then that means
they are supported.

I have one folder for each backend. When using the cusparse backend we
currently only test cusparse 12.8 so that is what I have started with,
but it is easy to extend this to support more (or all) cusparse versions
by creating additional folders, i.e. `clients/tests/cuda/12.0`,
`clients/tests/cuda/12.1`, `clients/tests/cuda/12.2`, ...etc.

This PR also make it trivial to partially support routines. For example,
in cusparse 12.8, only i32 index types are supported in spgemm and
spgemmreuse. Because each backend has its own yaml files, we can easily
specify using only i32 with cusparse 12.8 but both i32 and i64 (and
mixed i32/i64) with the rocm backend.

## Test Plan

Run all hipsparse tests as normal